refactor(fl/channels): relocate id_tracker and RxDevice impl into channels API
- Move `fl/id_tracker.{h,cpp.hpp}` -> `fl/channels/id_tracker.{h,cpp.hpp}`
  so the ID-mapping utility lives with its sole consumer (channel/strip
  tracking in ActiveStripData).
- Rename `fl/channels/rx/device.cpp.hpp` -> `fl/channels/rx.cpp.hpp` so the
  RxDevice factory implementation is the companion of `fl/channels/rx.h`
  where the class is declared (fixes CppHppHeaderPairChecker).
- Move `tests/fl/rx_device.cpp` -> `tests/fl/channels/rx.cpp` to mirror the
  source layout (fixes TestPathStructureChecker).
- Update unity-build headers accordingly and resolve the now-sorted
  alphabetical order in `fl/channels/rx/_build.cpp.hpp`.
- `ci/util/fbuild_runner.py`: return `shutil.which("fbuild")` directly;
  `os.fspath` wrapper tripped `ty`'s overload check and was redundant
  since `shutil.which` already returns `str | None`.
