+++ src/fl/channels/_build.cpp.hpp
@@ -8,7 +8,9 @@
 #include "fl/channels/config.cpp.hpp"
 #include "fl/channels/data.cpp.hpp"
 #include "fl/channels/driver.cpp.hpp"
+#include "fl/channels/id_tracker.cpp.hpp"
 #include "fl/channels/manager.cpp.hpp"
+#include "fl/channels/rx.cpp.hpp"
 #include "fl/channels/validation.cpp.hpp"
 #include "fl/channels/wave3.cpp.hpp"
 #include "fl/channels/wave8.cpp.hpp"
