+++ src/fl/channels/id_tracker.cpp.hpp
@@ -1,4 +1,4 @@
-#include "fl/id_tracker.h"
+#include "fl/channels/id_tracker.h"
 #include "fl/stl/noexcept.h"
 
 namespace fl {
@@ -7,10 +7,10 @@ int IdTracker::getOrCreateId(void* ptr) FL_NOEXCEPT {
     if (!ptr) {
         return -1;  // Invalid pointer gets invalid ID
     }
-    
+
     // Lock for thread safety
     mMutex.lock();
-    
+
     // Check if ID already exists
     auto it = mPointerToId.find(ptr);
     if (it != mPointerToId.end()) {
@@ -18,11 +18,11 @@ int IdTracker::getOrCreateId(void* ptr) FL_NOEXCEPT {
         mMutex.unlock();
         return id;
     }
-    
+
     // Create new ID
     int newId = mNextId++;
     mPointerToId[ptr] = newId;
-    
+
     mMutex.unlock();
     return newId;
 }
@@ -31,16 +31,16 @@ bool IdTracker::getId(void* ptr, int* outId) FL_NOEXCEPT {
     if (!ptr || !outId) {
         return false;
     }
-    
+
     // Lock for thread safety
     mMutex.lock();
-    
+
     auto it = mPointerToId.find(ptr);
     bool found = (it != mPointerToId.end());
     if (found) {
         *outId = it->second;
     }
-    
+
     mMutex.unlock();
     return found;
 }
@@ -49,12 +49,12 @@ bool IdTracker::removeId(void* ptr) FL_NOEXCEPT {
     if (!ptr) {
         return false;
     }
-    
+
     // Lock for thread safety
     mMutex.lock();
-    
+
     bool removed = mPointerToId.erase(ptr);
-    
+
     mMutex.unlock();
     return removed;
 }
@@ -62,9 +62,9 @@ bool IdTracker::removeId(void* ptr) FL_NOEXCEPT {
 size_t IdTracker::size() FL_NOEXCEPT {
     // Lock for thread safety
     mMutex.lock();
-    
+
     size_t currentSize = mPointerToId.size();
-    
+
     mMutex.unlock();
     return currentSize;
 }
@@ -72,10 +72,10 @@ size_t IdTracker::size() FL_NOEXCEPT {
 void IdTracker::clear() FL_NOEXCEPT {
     // Lock for thread safety
     mMutex.lock();
-    
+
     mPointerToId.clear();
     mNextId = 0;  // Reset ID counter to start at 0
-    
+
     mMutex.unlock();
 }
