+++ src/fl/channels/rx/_build.cpp.hpp
@@ -1,5 +1,4 @@
 /// @file _build.hpp
 /// @brief Unity build header for fl/channels/rx directory
 
-#include "fl/channels/rx/device.cpp.hpp"
 #include "fl/channels/rx/channel.cpp.hpp"
